+++ src/com/cpsc304/model/Pickup.java
@@ -8,7 +8,7 @@ public class Pickup extends Order {
 
     private Time readyTime;
 
-    public Pickup(Customer customer, int orderID, Date date, Time time, double amount, OrderStatus status, Restaurant restOrderedAt, Map<Food, Integer> quantity, Time readyTime) {
+    public Pickup(Customer customer, Long orderID, Date date, Time time, double amount, OrderStatus status, Restaurant restOrderedAt, Map<Food, Integer> quantity, Time readyTime) {
         super(customer, orderID, date, time, amount, status, restOrderedAt, quantity);
         this.readyTime = readyTime;
     }
